Fairly significant: pick, load, save, fcall, etc.
Added builtins, fcall, fcallargs to enable getting the 'range' function
"clear builtins range index list 5 int list print fcallargs"
For some reason, builtins appears on the stack as a dictionary, which
won't work with 'call' or 'callargs', so 'fcall', 'fcallargs', and 'sindex'
were created.

Added sindex to allow accessing Python dictionaries with string indexes.

Fixed pick command, was only returning top of stack.

Reworked 'int' command so that it returns a single value, not list, if there
is a single string without a comma. This allows the creation of a single
value on the stack (motivated in this case to enable 'index' to work well
with lists or dictionaries).
": range int list builtins range index list swap print fcallargs ;"

Reworked 'float', 'index', 'mm', 'mil', 'mils' similarly to int.

Added quoted string using "double quotes". All spaces inside the quote
marks are retained. Words are split on the double quote mark such that the 
following are equal:
1 2 3 " 4 5 6 " 7 8 9
1 2 3" 4 5 6 "7 8 9
If in a file, pairs of quote marks must be on the same line.

Added load and save commands for the user dictionary. Lightly tested.